Backend/requirements.txt
python-dotenv~=1.0.1
influxdb-client==1.43.0
fastapi~=0.111.0
uvicorn~=0.30.0
pydantic~=2.7.2